Create manifestUtils.ts
