+++ tools/botskills/src/utils/index.ts
@@ -9,3 +9,4 @@ export { ChildProcessUtils } from './childProcessUtils';
 export { getDispatchNames } from './dispatchUtils';
 export { sanitizePath, wrapPathWithQuotes } from './sanitizationUtils';
 export { isValidCultures, validatePairOfArgs, manifestV1Validation, manifestV2Validation } from './validationUtils';
+export { ManifestUtils } from './manifestUtils';
